tailortalk-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
supabase==2.9.1
httpx==0.27.2
PyJWT==2.8.0